working single pendulum with drag.
I feel like the drag is a bit off - there is probably a air component
and a friction component - I've tried modeling the air component as
depenent on v^2 and the friction component as dependeont on v but that's
probably not right. I think the weirdness shows up with very short
strings but I am not sure - maybe real pendulums behave that way too.